+++ src/main/java/gr/hua/dit/ds/reference/letter/service/repository/TeacherRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path="teachers")
-public interface TeacherRepository extends JpaRepository<Teacher, Long>{
+public interface TeacherRepository extends JpaRepository<Teacher, Integer>{
     Teacher findTeacherByUser(User user);
 }
